refactor(viewer): lift the symbolic-annotation walk out of the hook (step 1 of 2 for #2183)
useSymbolicAnnotations does the same whole-source parse that #2242 moved off
the main thread for grids and alignments: it hands the entire file to
parseSymbolicRepresentations, which decodes it to a JS string and grows a
main-thread WASM heap that never shrinks. Its guard is
hasEntityType(store, 'IfcAnnotation', 'IfcGridAxis'), so any model carrying a
grid trips it too — which is why fixing grids alone did not close the hole.

A worker module cannot import a React hook file, so the walk has to come out
first. This is that move and nothing else: the body from ensureBucket to
return result is byte-identical (verified by diff), every .free() keeps its
try/finally, and the moved types and helpers are re-exported from the hook so
no consumer changes.

The one thing that could not move is the hasEntityType pre-filter, which needs
store.entityIndex; it stays at the call site, ordered so the empty-source case
still reaches the original log.

Step 2 wires this into the overlay-parse worker and lands once #2242 is in.
Separating them keeps this PR a pure move and keeps step 2 off a non-main
base, since test.yml only runs for PRs targeting main.
